Fixes exercises 2-3 and 2-1!
- Fixes printing format for exercise 2-1 (the output was different from
expected)
- Fixes exercise 2-3 behavior

+++ C/2/3.c
@@ -8,15 +8,15 @@
  * 0x or 0X) into its equivalent integer value. The allowable digits are 0 through
  * 9, a through f, and A through F.
  * 
- * @param char s[]          An array of hexadecimal digits
+ * @param char s[]          A string of hexadecimal digits
  * 
- * @return long double      Integer result of operation
+ * @return long int         An integer number resultant from conversion
  */
 long int htoi(char s[])
 {
     int i = 0;
     int _size = strlen(s);
-    long double result = 0;
+    long int result = 0;
 
     /* Dealing with leading '0x' */
     if (_size > 2 && s[0] == '0' && (s[1] == 'x' || s[1] == 'X')) {
@@ -42,7 +42,7 @@ long int htoi(char s[])
             result = -1;
             break;
         }
-        result += value * pow(16, base);
+        result += ((long int) value) * ((long int) pow(16, base));
     }
     
     return result;
@@ -60,11 +60,11 @@ main ()
     char hexE[] = "ffffffffff";
     char hexF[] = "0X0a";
 
-    printf("%s : %.2f\n", hexA, htoi(hexA));
-    printf("%s : %.2f\n", hexB, htoi(hexB));
-    printf("%s : %.2f\n", hexC, htoi(hexC));
-    printf("%s : %.2f\n", hexD, htoi(hexD));
-    printf("%s : %.2f\n", hexE, htoi(hexE));
-    printf("%s : %.2f\n", hexF, htoi(hexF));
+    printf("%s : %li\n", hexA, htoi(hexA));
+    printf("%s : %li\n", hexB, htoi(hexB));
+    printf("%s : %li\n", hexC, htoi(hexC));
+    printf("%s : %li\n", hexD, htoi(hexD));
+    printf("%s : %li\n", hexE, htoi(hexE));
+    printf("%s : %li\n", hexF, htoi(hexF));
 }
